# src/main/resources/openapi/api.yml
openapi: 3.0.3
info:
  title: Proyecto 3 BC43 Grupo 5 - Leonardo Condori Ancco
  description: Proyecto 1 <br> El sistema a desarrollar está planteado en el contexto del negocio bancario que a medida que se va avanzando en los proyectos, se irá ampliando en base a este mismo proyecto. <br><br> Proyecto 2 <br> El sistema a desarrollar está planteado en el contexto del negocio bancario que extiende las funcionalidades y requerimientos presentados en el proyecto I. Por lo tanto, en este enunciado solo se agregan nuevas características o modificaciones a las ya presentada en el proyecto anterior. <br><br> Proyecto 3 <br> El sistema a desarrollar está planteado en el contexto del negocio bancario que extiende las funcionalidades y requerimientos presentados en el proyecto II. Por lo tanto, en este enunciado solo se agregan nuevas características o modificaciones a las ya presentada en el proyecto anterior.
  version: 1.0.0
...
paths:
  /clients:
    get:
      tags:
        - Clientes
      summary: Obtener todos los clientes
      operationId: getAllClients
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
    post:
      tags:
        - Clientes
      summary: Crear un nuevo cliente
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '201':
          description: Cliente creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
  /clients/{id}:
    get:
      tags:
        - Clientes
      summary: Obtener un cliente por su ID
      operationId: getClientById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: Cliente no encontrado
    put:
      tags:
        - Clientes
      summary: Actualizar un cliente por su ID
      operationId: updateClientById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '200':
          description: Cliente actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: Cliente no encontrado
    delete:
      tags:
        - Clientes
      summary: Eliminar un cliente por su ID
      operationId: deleteClientById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Cliente eliminado exitosamente
...
components:
  schemas:
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/ClientType'
    ClientType:
      type: string
      enum:
        - PERSONAL
        - BUSINESS
        - PERSONAL_VIP
        - BUSINESS_PYME
...